adding armor-related tests. handling crash of tail thread. add armor logging to history

+++ heathskin/tail_thread.py
@@ -1,6 +1,7 @@
 import threading
 import logging
 import subprocess
+import time
 
 
 class TailThread(threading.Thread):
@@ -12,17 +13,32 @@ class TailThread(threading.Thread):
         self.hs_log_path = hs_log_path
         self.queue = queue
         self.stop_flag = stop_flag
+        self.stopped = False
 
         self.process = None
 
     def run(self):
         self.logger.info("Starting log tailer at path %s", self.hs_log_path)
-        self.process = subprocess.Popen(
-            ["tail", "-f", "-c", "0", self.hs_log_path], stdout=subprocess.PIPE)
+        self.process = subprocess.Popen([
+            "tail",
+            "-f",
+            "-c", "0",
+            self.hs_log_path],
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE)
+
+        # XXX: yes, we need to sleep at least briefly
+        time.sleep(0.5)
+        self.process.poll()
 
         try:
             while True:
                 if self.stop_flag.is_set():
+                    self.shutdown()
+                    return
+                if self.process.returncode:
+                    e = "".join(self.process.stderr.readlines()).rstrip()
+                    self.logger.error("tail process crashed:\n%s", e)
                     self.shutdown()
                     return
 
@@ -40,6 +56,8 @@ class TailThread(threading.Thread):
 
     def shutdown(self):
         self.logger.info("Shutting down log tailer...")
-        self.process.kill()
-        self.process.wait()
+        if not self.process.returncode:
+            self.process.kill()
+            self.process.wait()
+        self.stopped = True
         self.logger.info("Log tailer shutdown")
